idenso

Macro color_f

macro_rules! color_f {
    ([$($a:expr),+ $(,)?], $($rest:tt)+) => { ... };
    ($base:ident . $a:ident, $($rest:tt)+) => { ... };
    ($a:expr, $($rest:tt)+) => { ... };
    (@tensor $a:expr; [$($b:expr),+ $(,)?], $($rest:tt)*) => { ... };
    (@tensor $a:expr; $base:ident . $b:ident, $($rest:tt)*) => { ... };
    (@tensor $a:expr; $b:expr, $($rest:tt)*) => { ... };
    (@tensor2 $a:expr, $b:expr; [$($c:expr),+ $(,)?] $(,)?) => { ... };
    (@tensor2 $a:expr, $b:expr; $base:ident . $c:ident $(,)?) => { ... };
    (@tensor2 $a:expr, $b:expr; $c:expr $(,)?) => { ... };
    (@tensor_done $a:expr, $b:expr, $c:expr) => { ... };
}

Builds an adjoint color structure-constant atom f(a,b,c).

Arguments can be typed slots, atoms, atom views, pattern variables such as RS.a__, or bracketed adjoint representation argument lists such as [CS.adj_, RS.a_]. Bare Rust identifiers and literals are expressions, so local slot bindings and numeric labels are used as-is.

Examples

use idenso::color_f;
use spenso::slot;

let expr = color_f!(slot!(coad_na, a), slot!(coad_na, b), slot!(coad_na, c));
let numeric_label_tensor = color_f!(1, 2, 3);
let pattern_tensor = color_f!(RS.a__, RS.b__, RS.c__);
let dimensioned_pattern = color_f!([CS.adj_, RS.a_], [CS.adj_, RS.b_], [CS.adj_, RS.c_]);