idenso

Macro color_t

macro_rules! color_t {
    ([$($a:expr),+ $(,)?], $($rest:tt)+) => { ... };
    ($base:ident . $a:ident, $($rest:tt)+) => { ... };
    ($a:expr, $($rest:tt)+) => { ... };
    ([$($a:expr),+ $(,)?] $(,)?) => { ... };
    ($base:ident . $a:ident $(,)?) => { ... };
    ($a:expr $(,)?) => { ... };
    (@tensor $a:expr; [$($i:expr),+ $(,)?], $($rest:tt)*) => { ... };
    (@tensor $a:expr; $base:ident . $i:ident, $($rest:tt)*) => { ... };
    (@tensor $a:expr; $i:expr, $($rest:tt)*) => { ... };
    (@tensor2 $a:expr, $i:expr; [$($j:expr),+ $(,)?] $(,)?) => { ... };
    (@tensor2 $a:expr, $i:expr; $base:ident . $j:ident $(,)?) => { ... };
    (@tensor2 $a:expr, $i:expr; $j:expr $(,)?) => { ... };
    (@tensor_done $a:expr, $i:expr, $j:expr) => { ... };
}

Builds a color-generator atom.

With one argument, this builds a fundamental generator factor for use inside chain! or trace!. The factor uses the chain placeholder indices in and out; the surrounding chain or trace owns the physical endpoints.

With three arguments, this builds an explicit generator t(a,i,j). The first argument is an adjoint color index, the second is a fundamental color index, and the third is an anti-fundamental color index.

Arguments can be typed slots, atoms, atom views, pattern variables such as RS.a__, or bracketed representation argument lists such as [CS.adj_, RS.a_]. Bare Rust identifiers and literals are expressions, so local slot bindings and numeric labels are used as-is.

Examples

use idenso::color_t;
use spenso::{slot, trace};

let factor = color_t!(slot!(coad_na, a));
let expr = trace!(&cof_nc, factor);
let factor_from_binding = color_t!(slot!(coad_na, a));
let explicit_tensor = color_t!(slot!(coad_na, a), slot!(cof_nc, i), slot!(cof_nc.dual(), j));
let pattern_tensor = color_t!(RS.a__, RS.i__, RS.j__);
let dimensioned_pattern = color_t!([CS.adj_, RS.a_], [CS.nc_, RS.i_], [CS.nc_, RS.j_]);