idenso

Macro u

macro_rules! u {
    ($label:expr, $base:ident . $i:ident) => { ... };
    ($label:expr, [$($i:expr),+ $(,)?]) => { ... };
    ($label:expr, $i:ident) => { ... };
    ($label:expr, $i:literal) => { ... };
    ($label:expr, $i:expr) => { ... };
    (@done $label:expr, $i:expr) => { ... };
}

Builds a Dirac u spinor tensor.

The first argument is the spinor label. The second argument can be a typed bispinor slot, an atom-like expression, an identifier/literal that defaults to a four-dimensional bispinor slot, or a bracketed bispinor argument list such as [RS.d_, RS.i_].

Examples

use idenso::u;
use spenso::slot;

let default_spinor = u!(1, i);
let indexed_spinor = u!(1, 2);
let explicit_spinor = u!(1, slot!(bis_d, i));
let pattern_spinor = u!(1, [RS.d_, RS.i_]);