idenso::shorthands::chain

Trait Chain

pub trait Chain {
    // Required methods
    fn collect_chains(&self, representation: LibraryRep) -> Atom;
    fn normalize_chains(&self) -> Atom;
    fn undo_single_length(&self) -> Atom;
    fn chainify(&self, representation: LibraryRep) -> Atom;
}

Required Methods

fn collect_chains(&self, representation: LibraryRep) -> Atom

combine adjacent chain expressions into a single chain expression.

chain(rep(d,i), rep(d,j), ...,A(..,in,out...))*chain(rep(d,j), rep(d,k), B(..,in,out...),...)

becomes

chain(rep(d,i), rep(d,k), ...,A(..,in,out...),B(..,in,out...),...)

fn normalize_chains(&self) -> Atom

Turns traced out chains into traces e.g.

chain(rep(d,i), rep(d,i), ...) becomes trace(rep(d),...)

fn undo_single_length(&self) -> Atom

and single length chains back into the corresponding tensor

fn chainify(&self, representation: LibraryRep) -> Atom

turns tensors with two indices of the representation into chain expressions, for collecting using [Atom::collect_chains]

A(..,rep(d,i),rep(d,j),...) becomes chain(rep(d,i),rep(d,j),A(..,in,out...))