idenso::tensor

Struct SymbolicTensor

pub struct SymbolicTensor<Aind: AbsInd = AbstractIndex> {
    pub structure: OrderedStructure<LibraryRep, Aind>,
    pub is_metric: bool,
    pub is_composite: bool,
    pub expression: Atom,
}

A fully symbolic tensor, with no concrete values.

This tensor is used to represent the structure of a tensor, and is used to perform symbolic contraction. Currently contraction is just a multiplication of the atoms, but in the future this will ensure that internal indices are independent accross the contraction.

It can also serve as a tensor structure while retaining the symbolic expression and the ordered external-index structure across transformations.

Fields
structure: OrderedStructure<LibraryRep, Aind>is_metric: boolis_composite: boolexpression: Atom
Implementations

impl<Aind: AbsInd> SymbolicTensor<Aind>

pub fn from_named<N>(structure: &N) -> Option<Self>
where N: ToSymbolic + HasName + TensorStructure<Slot = LibrarySlot<Aind>>, N::Name: IntoSymbol + Clone, N::Args: IntoArgs,

pub fn from_permuted<N>(structure: &PermutedStructure<N>) -> Option<Self>
where N: ToSymbolic + HasName + TensorStructure<Slot = LibrarySlot<Aind>>, N::Name: IntoSymbol + Clone, N::Args: IntoArgs,

pub fn to_named(&self) -> NamedStructure<Symbol, Vec<Atom>, LibraryRep, Aind>

pub fn empty(expression: Atom) -> Self

pub fn get_atom(&self) -> &Atom

pub fn to_network( &self, library: &TensorLibrary<MixedTensor<f64, ExplicitKey<AbstractIndex>>, AbstractIndex>, ) -> Result<Network<NetworkStore<MixedTensor<f64, ShadowedStructure<AbstractIndex>>, Atom>, ExplicitKey<AbstractIndex>, Symbol, AbstractIndex>, TensorNetworkError<ExplicitKey<AbstractIndex>, Symbol>>