idenso::tensor

Type Alias SymbolicNet

pub type SymbolicNet<Aind> = Network<NetworkStore<SymbolicTensor<Aind>, Atom>, DummyKey, Symbol, Aind>;

Aliased Type

pub struct SymbolicNet<Aind> {
    pub graph: NetworkGraph<DummyKey, Symbol, Aind>,
    pub store: NetworkStore<SymbolicTensor<Aind>, Atom>,
    pub state: NetworkState,
}

Fields
graph: NetworkGraph<DummyKey, Symbol, Aind>

Expression and contraction topology whose leaves reference stored values.

store: NetworkStore<SymbolicTensor<Aind>, Atom>

Backing tensor and scalar storage referenced by graph leaves.

state: NetworkState

Classification of the network result as scalar or tensor, including self-duality.